1 - Get data by bodyType (moon) from the Solar System API
First thing, we need to retrieve data from the Solar System API from which we want to build our Documentation Website. We have filtered the data we want to retrieve by body type.

Indeed, in our next commit, we will work on implementing dynamic urls based on the englishName property of each object from the response array.
Then we'll make a new request to the Wikipedia API to try and get associated images.