2.1 - WIP : Implement dynamic wikipedia URLs to get the thumbnail image of each object
In this part, we have retrieved the thumbnail URl of each object from the Wikipedia API.

However, many objects don't have an associated image.

We have to find out if it's because there is none or if its bc of the custom name we've made with the addition of ' (moon)' in our dynamicWikiUrlsArray (ie. lines 23 - 26).